feat: タスクの残っているグループを削除しようとした時に例外を投げるように

+++ app/Contents/TaskGroup/TaskGroupDestroyService.php
@@ -3,11 +3,16 @@
 namespace App\Contents\TaskGroup;
 
 use App\Models\TaskGroup;
+use App\Exceptions\TodoException;
 
 class TaskGroupDestroyService
 {
     public function execute(TaskGroup $taskGroup): void
-    {  
+    {
+        if ($taskGroup->tasks()->exists()) {
+            throw new TodoException('TE00002');
+        }
+
         $taskGroup->delete();
     }
 }
